requirements.txt
torch
torchvision
scikit-learn
seaborn
matplotlib
numpy
opencv-python
Pillow
streamlit
